+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
@@ -383,7 +383,7 @@ public class User implements Serializable {
 		return selectedCharts;
 	}
 
-    public Byte getSelectedChartId(ChartType chartType) {
+    public Integer getSelectedChartId(ChartType chartType) {
         for(Chart chart: selectedCharts)
             if(chartType.equals(chart.getType()))
                 return chart.getI();
