+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/ChartRepository.java
@@ -1,18 +1,17 @@
 package mobi.nowtechnologies.server.persistence.repository;
 
-import java.util.List;
-
 import mobi.nowtechnologies.server.persistence.domain.Chart;
 import mobi.nowtechnologies.server.shared.enums.ChartType;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
 
 /**
  * @author Alexander Kolpakov (akolpakov)
  *
  */
-public interface ChartRepository extends JpaRepository<Chart, Byte> {
+public interface ChartRepository extends JpaRepository<Chart, Integer> {
 
 	@Query("select chart from Chart chart join chart.communities community where community.rewriteUrlParameter like ?1 order by chart.name asc")
 	List<Chart> getByCommunityURL(String communityURL);
